requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
argcomplete==3.6.3
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==4.4.2
-e git+https://github.com/ByteDance-Seed/depth-anything-3@2c21ea849ceec7b469a3e62ea0c0e270afc3281a#egg=depth_anything_3
distlib==0.4.0
e3nn==0.6.0
einops==0.8.2
evo==1.34.3
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
identify==2.6.17
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.5.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.6.1
nodeenv==1.10.0
numexpr==2.14.1
numpy==2.4.3
omegaconf==2.3.0
opencv-python==4.13.0.92
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pillow_heif==1.3.0
platformdirs==4.9.4
plyfile==1.1.3
polars==1.39.0
polars-runtime-32==1.39.0
pre_commit==4.5.1
proglog==0.1.12
psutil==7.2.2
pycolmap==3.13.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-discovery==1.1.3
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rosbags==0.11.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scipy==1.17.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
thop==0.1.1.post2209072238
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.3
transformers==5.3.0
trimesh==4.11.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.2.0
xformers==0.0.35
zstandard==0.25.0
